drivers/net/aspeednic: Fix RX buffer size
The hardware defaults to 1600 bytes of RX buffer size, currently we
allocate a little less and there is a chance that the hardware might DMA
more than we've allocated.

In the process of writing this patch, it was discovered that the driver
was incorrectly informing the hardware of its RX buffer size, addressed
that problem too.

Signed-off-by: Cyril Bur <cyril.bur@au1.ibm.com>
Signed-off-by: Joel Stanley <joel@jms.id.au>