Add pinctrl_rgmii4_default phandle to pinctrl-0 in mac3 definition (#2)
Summary:
A follow up to the initial commit. This change is to fix on of the Caveats listed with that commit:

--Ping between U-boot and external server does not work after a power cycle. (Bug. Workaround: Allow OS to boot up and warm reboot to stop at U-boot to perform the ping).

The fix is to add pinctrl_rgmii4_default to pinctrl-0 in mac3 definition in the uboot dts file.

X-link: https://github.com/facebookexternal/openbmc.cisco/pull/2

Reviewed By: tao-ren

Differential Revision: D35015215

Pulled By: binhuang00

fbshipit-source-id: 8663c3ee63c4eec3dc6014e4f97fb272f843d484