enable the primary watchdog in u-boot for fblite
Summary:
# Description

Current platforms that use the Fblite configurations aren't utilizing the primary watchdog. This PR adds an implementation that supports starting the watchdog in uboot on those platforms.

A kconfig entry is added for CONFIG_ASPEED_ENABLE_PRIMARY_WATCHDOG (defaults to disabled). This config is enabled for all Fblite platforms. When enabled, the primary watchdog (wdt1) will start in uboot even when the dual boot watchdog (fmc_wdt) is also configured to start. When both watchdogs are enabled, wdt1 will have a higher timeout (2x that of fmc_wdt) that will later be updated to 30 seconds by the linux kernel.

Note that the wdt1 config is only considered when the fmc_wdt is enabled. In the case where fmc_wdt is disabled, wdt1 will start regardless of the wdt1 config value.

X-link: https://github.com/facebookexternal/openbmc.arista/pull/293

Reviewed By: Scott8440

Differential Revision: D48342286

Pulled By: tao-ren

fbshipit-source-id: 6c3d3490c76ff93f33138831e436dc77b3657e68